requirements.txt
requests
pydantic
pydantic-settings
pyjwt[crypto]
pyjks